feat(locale): use translated descriptions for startup argument search
